Corrigido erros de merge #ALL

+++ codigo/comidinhasVeg/src/main/java/javaquinho/comidinhas/models/Mesa.java
@@ -1,14 +1,10 @@
 package javaquinho.comidinhas.models;
 
-import com.fasterxml.jackson.annotation.JsonBackReference;
-
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
@@ -32,7 +28,7 @@ public class Mesa {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id", unique = true, nullable = false)
-    private Long idMesa;
+    private int idMesa;
 
     /**
      * Capacidade máxima de pessoas que a mesa pode acomodar.
